[feat] at_most_one, partition_map, sorted, sorted_by, sorted_by_cached_key sorted_by_key

+++ playground/examples/_array_as_it_fuse.rs
@@ -1,13 +1,13 @@
 fn main() {
     struct Foo {
-        counter: i32,
+        counter: isize,
     }
 
     // カウンタが偶数のときだけその値を返す
     impl Iterator for Foo {
-        type Item = i32;
+        type Item = isize;
 
-        fn next(&mut self) -> Option<i32> {
+        fn next(&mut self) -> Option<isize> {
             let val = self.counter;
             self.counter += 1;
             if val % 2 == 0 {
